Add files via upload
The program will prompt user to input name, pin code, transaction type (balance
enquiry(“show the balance on screen” or “show the balance on receipt”), deposit money(the
amount of money which you want to make deposit into your account), withdraw money(“the amount of money that you want to withdraw”).The program will output
different type of responds depending on enquiries. Amount of money in the bank accounts will
be constant integer and pin code of card will be provided by me. In conclusion ,at the end of
different types of process, program will output some type of expressions of completion and thank.